+++ tests/frontendIntegration/django2x/polls/views.py
@@ -436,6 +436,11 @@ def get_info(request: HttpRequest):
     return HttpResponse("")
 
 
+def check_rid_no_session(request: HttpRequest):
+    rid = request.headers.get("rid")  # type: ignore
+    return HttpResponse("fail" if rid is None else "success")
+
+
 @custom_decorator_for_update_jwt()
 @verify_session()
 def update_jwt(request: HttpRequest):