+++ tests/frontendIntegration/fastapi-server/app.py
@@ -319,6 +319,12 @@ async def get_info(r_session: SessionContainer = Depends(verify_session())):
     )
 
 
+@app.get("/check-rid-no-session")
+def check_rid_no_session_api(request: Request):
+    rid = request.headers.get("rid")
+    return PlainTextResponse("fail" if rid is None else "success")
+
+
 @app.options("/update-jwt")
 def update_options():
     return send_options_api_response()